+++ tests/test_helpers.py
@@ -340,3 +340,14 @@ def test_profile_parser_(value, expected):
 def test_tokenize_keywords(value, expected):
     actual = helpers.tokenize_keywords(value)
     assert actual == expected
+
+
+@pytest.mark.parametrize('url,expected', (
+    ('/foo/bar/', '/foo/bar/'),
+    ('/foo/bar/?page=2', '/foo/bar/'),
+    ('/foo/bar/?page=2&baz=3', '/foo/bar/?baz=3'),
+
+))
+def test_get_pagination_url(rf, url, expected):
+    request = rf.get(url)
+    assert helpers.get_pagination_url(request, 'page') == expected